+++ printing/android/src/main/java/net/nfet/flutter/printing/PrintingJob.java
@@ -86,7 +86,6 @@ public class PrintingJob extends PrintDocumentAdapter {
         result.put("directPrint", false);
         result.put("dynamicLayout", canPrint);
         result.put("canPrint", canPrint);
-        result.put("canConvertHtml", canRaster);
         result.put("canShare", true);
         result.put("canRaster", canRaster);
         return result;